+++ src/main/java/eu/bbmri_eric/negotiator/common/configuration/security/NegotiatorPrincipal.java
@@ -1,5 +1,6 @@
-package eu.bbmri_eric.negotiator.user;
+package eu.bbmri_eric.negotiator.common.configuration.security;
 
+import eu.bbmri_eric.negotiator.user.Person;
 import java.security.Principal;
 
 public interface NegotiatorPrincipal extends Principal {
