+++ src/main/java/eu/bbmri_eric/negotiator/negotiation/state_machine/negotiation/NegotiationLifecycleServiceImpl.java
@@ -1,9 +1,9 @@
 package eu.bbmri_eric.negotiator.negotiation.state_machine.negotiation;
 
+import eu.bbmri_eric.negotiator.common.configuration.security.NegotiatorUserDetailsService;
 import eu.bbmri_eric.negotiator.common.exceptions.EntityNotFoundException;
 import eu.bbmri_eric.negotiator.common.exceptions.WrongRequestException;
 import eu.bbmri_eric.negotiator.negotiation.NegotiationRepository;
-import eu.bbmri_eric.negotiator.user.NegotiatorUserDetailsService;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
